[tutorials] Fix and improve Dev. Env. Setup tutorial.
Remove `kytos-napps` from the development environment setup tutorial,
since now everyone should install them from the napps-server (except if
trying to change one of them, which should not be the case for our
tutorials). Also change the tutorial to consider the installation of the
projects (python-openflow, kytos-utils and kytos) requirements by the
setup script, without the need of manually installing them.